Separate out components used in Export
