+++ app/components/general/SourcesSelection/index.tsx
@@ -19,7 +19,7 @@ import {
     CheckboxProps,
 } from '@the-deep/deep-ui';
 import { EntriesAsList } from '@togglecorp/toggle-form';
-import { useQuery } from '@apollo/client';
+import { useQuery, gql } from '@apollo/client';
 
 import { organizationTitleSelector } from '#components/selections/NewOrganizationSelectInput';
 import SourcesFilter, { getProjectSourcesQueryVariables } from '#components/leadFilters/SourcesFilter';
@@ -30,8 +30,108 @@ import {
     ProjectSourceListQueryVariables,
 } from '#generated/types';
 import { isFiltered } from '#utils/common';
-import { PROJECT_LEADS } from '../queries';
 import styles from './styles.css';
+
+const PROJECT_LEADS = gql`
+    query ProjectSourceList(
+        $projectId: ID!,
+        $page: Int,
+        $pageSize: Int,
+        $ordering: [LeadOrderingEnum!],
+        $assignees: [ID!],
+        $createdBy: [ID!],
+        $authoringOrganizationTypes: [ID!],
+        $confidentiality: LeadConfidentialityEnum,
+        $createdAtGte: DateTime,
+        $createdAtLte: DateTime,
+        $emmEntities: String,
+        $emmKeywords: String,
+        $emmRiskFactors: String,
+        $priorities: [LeadPriorityEnum!],
+        $publishedOnGte: Date,
+        $publishedOnLte: Date,
+        $search: String,
+        $statuses: [LeadStatusEnum!],
+        $sourceOrganizations: [ID!],
+        $authorOrganizations: [ID!],
+        $entriesFilterData: EntriesFilterDataInputType,
+        $hasEntries: Boolean,
+        $hasAssessment: Boolean,
+    ) {
+        project(id: $projectId) {
+            id
+            leads (
+                page: $page,
+                pageSize: $pageSize,
+                ordering: $ordering,
+                assignees: $assignees,
+                createdBy: $createdBy,
+                authoringOrganizationTypes: $authoringOrganizationTypes,
+                confidentiality: $confidentiality,
+                createdAtGte: $createdAtGte,
+                createdAtLte: $createdAtLte,
+                emmEntities: $emmEntities,
+                emmKeywords: $emmKeywords,
+                emmRiskFactors: $emmRiskFactors,
+                priorities: $priorities,
+                publishedOnGte: $publishedOnGte,
+                publishedOnLte: $publishedOnLte,
+                search: $search,
+                statuses: $statuses,
+                sourceOrganizations: $sourceOrganizations,
+                authorOrganizations: $authorOrganizations,
+                entriesFilterData: $entriesFilterData,
+                hasEntries: $hasEntries,
+                hasAssessment: $hasAssessment,
+            ) {
+                totalCount
+                page
+                pageSize
+                results {
+                    id
+                    clientId
+                    createdAt
+                    title
+                    publishedOn
+                    createdBy {
+                        id
+                        displayName
+                    }
+                    project
+                    authors {
+                        id
+                        title
+                        mergedAs {
+                            id
+                            title
+                        }
+                    }
+                    assignee {
+                        id
+                        displayName
+                    }
+                    source {
+                        mergedAs {
+                            id
+                            title
+                        }
+                        id
+                        url
+                        title
+                    }
+                    entriesCount {
+                        total
+                    }
+                    filteredEntriesCount
+                    leadPreview {
+                        pageCount
+                    }
+                    isAssessmentLead
+                }
+            }
+        }
+    }
+`;
 
 type Project = NonNullable<ProjectSourceListQuery['project']>;
 type Lead = NonNullable<NonNullable<NonNullable<Project['leads']>['results']>[number]>;
