+++ app/views/Sources/index.tsx
@@ -216,6 +216,7 @@ function Sources(props: Props) {
             }
             const {
                 // NOTE: we are not using ordering
+                // eslint-disable-next-line @typescript-eslint/no-unused-vars
                 ordering: orderingFilter,
                 ...others
             } = userSavedLeadFilter.filters;
